+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Remove runtime dependency on [Fkit](https://github.com/nullobject/fkit)
 * Mark subscriptions as closed when an observer unsubscribes
 * Allow `Signal.of` to take multiple arguments
+* Don't allow functions which take multiple arguments to an array (use the array spread syntax instead)
 
 ## 4.0.1 (2019-02-04)